+++ create_table.h
@@ -6,10 +6,10 @@
 /**
 * create a shared device status table
 */
-float* create_status_table(key_t key);
+int* create_status_table(key_t key);
 
 /**
 * create user mode table
 */
-float* create_mode_table(key_t key);
+int* create_mode_table(key_t key);
 # endif
